requirements.txt
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
watchdog==3.0.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0